generator controller
movement based on gamemanager calls
