--- Assets/GeneratorController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f02d74ecf26c211449c07d218be86e98
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
